changes: Add RemoveAttention
